requirements.txt
pandas>=1.3.0
kafka-python>=2.0.2
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-cov>=2.12.1
flake8>=3.9.2
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Cors>=3.0.10
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.3.0
apache-airflow>=2.7.0
apache-airflow-providers-apache-kafka>=1.1.0
apache-airflow-providers-postgres>=5.7.0 